Do not flush the database when using bok-choy in testsonly mode.
Testsonly will not work currently, because the database flush will
remove some data fixtures. Even when they are loaded at the beginning
of the next test run, there are database corruptions. Instead of
flushing, do not do anything with the database at the end of the
testsonly test run. That way, tests that use dynamic data can be
run many times in a row.